# API.yaml
...
swagger: "2.0"
info:
  description: "Calculates minimum path between points depending on user generated\
    \ map input."
  version: "1.0.0"
  title: "Team 2 Pathfinding API"
host: "ec2-54-200-185-55.us-west-2.compute.amazonaws.com:8080"
basePath: "/v2"
schemes:
- "http"
paths:
...
  /id={problem_id}/Goal:
    get:
      tags:
      - "Goal"
      summary: "Goal Location"
      description: "Returns a description of the goal location.\n"
      operationId: "get_goal"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "The goal location information"
          schema:
            $ref: "#/definitions/Goal"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.goal_controller"
    put:
      tags:
      - "Goal"
      summary: "Update the existing goal value"
      description: ""
      operationId: "update_goal"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "goal"
        description: "Goal object that needs to be updated."
        required: true
        schema:
          $ref: "#/definitions/Goal"
      responses:
        403:
          description: "ProblemID busy"
        404:
          description: "Goal not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.goal_controller"
  /id={problem_id}/Boundary:
    get:
      tags:
      - "Boundary"
      summary: "Boundary"
      description: "Returns a description of the boundary"
      operationId: "get_boundary"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Boundary information, including starting coordinates, width\
            \ and height"
          schema:
            $ref: "#/definitions/Boundary"
        403:
          description: "ProblemID busy"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "swagger_server.controllers.boundary_controller"
    put:
      tags:
      - "Boundary"
      summary: "Update the existing boundary value"
      description: ""
      operationId: "update_boundary"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "problem_id"
        in: "path"
        description: "The id of the problem being manipulated"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "boundary"
        description: "Boundary object that needs to be updated."
        required: true
        schema:
          $ref: "#/definitions/Boundary"
      responses:
        403:
          description: "ProblemID busy"
        404:
          description: "Boundary not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.boundary_controller"
...
definitions:
  Coordinate:
    type: "object"
    properties:
      longitude:
        type: "number"
        format: "double"
        description: "the longitude of the object."
      latitude:
        type: "number"
        format: "double"
        description: "the latitude of the object."
  Rectangle:
    type: "object"
    properties:
      coordinates:
        $ref: "#/definitions/Coordinate"
      width:
        type: "number"
        format: "double"
        description: "The width of the rectangle"
      height:
        type: "number"
        format: "double"
        description: "The height of the rectangle"
...
  Goal:
    type: "object"
    properties:
      coordinates:
        $ref: "#/definitions/Coordinate"
  Boundary:
    type: "object"
    properties:
      boundary_info:
        $ref: "#/definitions/Rectangle"
...
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"